requirements.txt
insightface>=0.7
onnxruntime-gpu
onnx
opencv-python
numpy
aiortc
av
websockets
pyvirtualcam
